Brief
Pooja Magic Breeze is one of the most exciting projects by Pooja Ventures Pvt. Ltd. Build by one of Hyderabad's top builders. The project is strategically located in Narsingi, Hyderabad, and offers luxuriously built 3BHK Sky Villas for sale in Hyderabad ranging in size from 2600 to 3634 sq. ft. These villas are suitable for anyone looking for more than just a place to live; perfect for nature lovers, you will be able to relax in a tranquil environment. 
The villas are well-equipped with modern amenities that are needed for comfortable and hassle-free living and are ideal for a high-end lifestyle. Pooja magic breeze also has high-quality fittings and features. The well-designed floor plan provides you with enough space. The villas are well-ventilated and have a gentle light flow. Each of these villas has double-height gardens and sit-out areas.
•	Property Name:    Pooja Magic Breeze
•	Property Type:      Sky Villa
•	Locality:                  Narsinghi, Hyderabad
•	Bedrooms:             3BHK Duplex Flats
•	Total Acres:           3.8 acres	
•	Total Units:           166 Units
•	Total Towers:       3 Towers
•	Total Floors:         G+10 floors
•	Project Status:     Under Construction
•	Possession Date: October 2022
•	Price per sqft:       Rs. 6500/Price per sqft
•	Size Range:           2200 Sqft to 3600 Sqft
•	Price Range:         Rs. 1.6 Cr  to 2.5 Cr
•	RERA-registered: P02400000029 registration number
Builder Details:
The Magic Breeze Landscape was designed by landscape architecture firm Penda for developers Pooja Crafted Homes and is expected to begin construction later this year in Hyderabad, India. The architecture of Penda, which spans 8000 square metres, is inspired by India's stepwells and water mazes. Pooja Ventures is a well-known real estate company. It was established in 2001 and has since become known for its excellence in the creation of prime properties in cities such as Hyderabad, Bangalore, Mangalore, and Vijayawada. The community aims to remain current in the fields of construction and growth in order to provide you with the best service possible.
They are committed to satisfying the needs of their customers and offering unsurpassed efficiency. The community ensures that quality is a constant and mutual aim in all of their projects. They locate property opportunities and optimize their growth potential using strategic analysis and strong experience.
Unit Details:
From the Pooja venture, the Magic Breeze offers 3 BHK Duplex residential apartments. 3 BHK apartment sizes range from 2200 to 3600 square feet. The project is spread out over 3.8 acres of lush greenery area. It has total 116 units of 3 BHK Duplex model, It's uncommon to come across an architectural typology in which function and beauty are so inextricably connected, and where human needs and environmental impact are so well balanced.
Price Details:
Pooja Magic Breeze is a 3.8-acre property. There are 166 units in the project. The project is in the middle of its growth. The available flat units include 3 BHK Flats with a minimum price of Rs 1.6 Cr and a maximum price of Rs 2.50 Cr, with property sizes ranging from 2200 Sqft to 3600 Sqft.
Unit Type	Size	Basic Cost	Price per Sqft
3 BHK	2200 sqft	1.6 cr	6500/-
3 BHK	2715 sqft	1.92 cr	6500/-
3 BHK	2915 sqft	2.05 cr	6500/-
3 BHK	3209 sqft	2.24 cr	6500/-
3 BHK	3600 sqft	2.50 cr	6500/-

GST will be extra as applicable on the total flat value exclusive of corpus fund (6% on the total flat value) and also registration charges extra as per govt. prevailing norms.(6% on the total flat value).
Luxury Amenities
The space will serve as a communal garden for the residents of the 145 apartments in the building, with features such as bamboo coves, gardens of flowers, herbs, and grasses, water-displays, and seating areas; a place where they can visit with family and friends or simply commune with nature in a quiet space within what is a famously densely populated country.
1.	Sports Amenities: 
•	Gymnasium
•	Badminton Court(s)
•	Basketball
•	Jogging / Cycle Track
•	Swimming Pool
•	Kids' Play Areas / Sand Pits
•	Yoga Areas
2.	Convenience:
•	Power backup
•	Lift
•	24*7 Water supply
•	2 car parks per flat
3.	Safety:
•	24*7 Security 
•	Fire Fighting Systems
•	CCTV / Video Surveillance
•	Intercom Facility
4.	Leisure:
•	Party Hall
•	Clubhouse
•	Indoor Games
5.	Environment:
•	Rain Water Harvesting
•	Sewage Treatment Plant
•	Plantation With Drip Irrigation
Location Details:
Magic Breeze is a one-of-a-kind community located adjacent to the Neknampur Lake on a scenic 4-acre lakefront site.
Pooja Magic Breeze is in Narsingi, West Zone, and is well connected by major highways such as the Nehru Outer Ring Road in Hyderabad.